Alterando de Float para BigDecimal para não ter problemas com ponto flutuante.

+++ src/main/java/com/cvcorp/hotel/domain/entity/Price.java
@@ -3,10 +3,12 @@ package com.cvcorp.hotel.domain.entity;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 public class Price extends BaseDomain {
 
-    private final Float adult;
-    private final Float child;
+    private final BigDecimal adult;
+    private final BigDecimal child;
 }
